requirements.txt
flask==3.0.2
flask-cors==4.0.0
numpy==1.26.4
pandas==2.2.1
tensorflow==2.15.0
plotly==5.19.0
python-dotenv==1.0.1
scikit-learn==1.3.2
pytest==8.0.2
black==24.2.0
flake8==7.0.0
gunicorn==21.2.0 